+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 2022-10-07
 
 * Fixed wrong `form-select` class uses for Bootstrap 4 template selects: replaced them by `custom-select`
+* Fixed Column action still displays original column value with ->when(false) : #103
 
 ## [5.0.1](https://github.com/Okipa/laravel-table/compare/5.0.0...5.0.1)